refactor: move CPU discovery into src/hostcpu.rs
Give the /sys core-topology probe its own module beside hostgpu.rs:
math_core_count (the only public item) plus the pure parsing helpers
and their tests move out of serve.rs unchanged.
